+++ src/cpp/espectre_sources.cmake
@@ -21,6 +21,7 @@ set(ESPECTRE_RUNTIME_COMMON_SOURCES
     "${ESPECTRE_CPP_ROOT}/runtime/protocol_json.cpp"
     "${ESPECTRE_CPP_ROOT}/runtime/runtime_config_utils.cpp"
     "${ESPECTRE_CPP_ROOT}/runtime/runtime_diagnostics.cpp"
+    "${ESPECTRE_CPP_ROOT}/runtime/runtime_diagnostics_protocol.cpp"
     "${ESPECTRE_CPP_ROOT}/runtime/runtime_time.cpp"
     "${ESPECTRE_CPP_ROOT}/runtime/frontend_command_engine.cpp"
 )